refactor: 테스트 파일 및 tanstack-query hook들 위치 변경 및 테스트 파일 타입 에러 해결
