+++ apps/web/features/auction-detail/model/useAuctionBidState.ts
@@ -3,9 +3,9 @@ import { useCallback, useEffect, useRef, useState } from 'react';
 import { useAuthStore } from '@/shared/stores/auth';
 import { Bid } from '@/shared/types/db';
 
-import { useAuctionBid } from '@/shared/hooks/useAuctionBid';
-import { useAuctionDetail } from '@/shared/hooks/useAuctionDetail';
-import { useRealtimeBids } from '@/shared/hooks/useRealTimeBid';
+import { useAuctionBid } from '@/shared/api/client/auction/useAuctionBid';
+import { useAuctionDetail } from '@/shared/api/client/auction/useAuctionDetail';
+import { useRealtimeBids } from '@/shared/api/client/auction/useRealTimeBid';
 import { useBidCostHandlers } from './useBidCostHandlers';
 import { useSendBid } from './useSendBid';
