+++ apps/web/shared/api/client/auction/useAuctionDetail.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { fetchAuctionDetail } from '../api/server/auction/fetchAuctionDetail';
-import { AuctionDetail } from '../types/db';
+import { AuctionDetail } from '../../../types/db';
+import { fetchAuctionDetail } from '../../server/auction/fetchAuctionDetail';
 
 // auction, product, seller, bids 모두 db 타입을 활용해 정의
