+++ apps/web/shared/api/client/auction/useRealTimeBid.ts
@@ -1,8 +1,8 @@
 import { useEffect, useRef } from 'react';
 
-import { fetchBidderName } from '../api/server/auction/fetchBidderName';
-import { supabase } from '../lib/supabase/supabase';
-import { Bid } from '../types/db';
+import { supabase } from '@/shared/lib/supabase/supabase';
+import { Bid } from '@/shared/types/db';
+import { fetchBidderName } from '../../server/auction/fetchBidderName';
 
 export function useRealtimeBids(auctionId: string | undefined, onNewBid: (bid: Bid) => void) {
   const channelRef = useRef<ReturnType<typeof supabase.channel> | null>(null);
