+++ apps/web/vitest.shims.d.ts
@@ -1,2 +1,3 @@
 /// <reference types="vitest/globals" />
 /// <reference types="vitest/import-meta" />
+/// <reference types="@testing-library/jest-dom/vitest" />
